requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
email-validator==2.1.1
Flask-Migrate==4.0.7
Flask-Admin==1.6.1
pytest==8.0.0